# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tasche API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "tags": [
          "タスク"
        ],
        "summary": "Get Tasks",
        "description": "タスク一覧を取得する.",
        "operationId": "get_tasks_api_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "アーカイブ済みタスクを含める",
              "default": false,
              "title": "Include Archived"
            },
            "description": "アーカイブ済みタスクを含める"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TaskListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "タスク"
        ],
        "summary": "Create Task",
        "description": "新規タスクを作成する.",
        "operationId": "create_task_api_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TaskResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "put": {
        "tags": [
          "タスク"
        ],
        "summary": "Update Task",
        "description": "既存タスクを更新する.",
        "operationId": "update_task_api_tasks__task_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TaskResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "タスク"
        ],
        "summary": "Delete Task",
        "description": "タスクを削除（アーカイブ）する.",
        "operationId": "delete_task_api_tasks__task_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TaskResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weeks/current": {
      "get": {
        "tags": [
          "週"
        ],
        "summary": "Get Current Week",
        "description": "現在の週情報を取得する.",
        "operationId": "get_current_week_api_weeks_current_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "timezone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "タイムゾーン（デフォルト: ユーザー設定値）",
              "title": "Timezone"
            },
            "description": "タイムゾーン（デフォルト: ユーザー設定値）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_WeekResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "週"
        ],
        "summary": "Update Current Week",
        "description": "現在の週設定を更新する.",
        "operationId": "update_current_week_api_weeks_current_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeekUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_WeekResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_TaskListResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskListResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[TaskListResponse]"
      },
      "APIResponse_TaskResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[TaskResponse]"
      },
...
      "APIResponse_WeekResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WeekResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[WeekResponse]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "タスク名"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TaskCreate",
        "description": "タスク作成リクエスト."
      },
      "TaskListResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Tasks",
            "description": "タスク一覧"
          }
        },
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "TaskListResponse",
        "description": "タスク一覧レスポンス."
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "タスクID（ULID形式、prefix: tsk_）"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "タスク名"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived",
            "description": "アーカイブフラグ"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時（ISO 8601）"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新日時（ISO 8601）"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "is_archived",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse",
        "description": "タスクレスポンス."
      },
      "TaskUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "タスク名"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TaskUpdate",
        "description": "タスク更新リクエスト."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WeekResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "週ID（ULID形式、prefix: wk_）"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "ユーザーID"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date",
            "description": "週開始日（YYYY-MM-DD）"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date",
            "description": "週終了日（YYYY-MM-DD）"
          },
          "unit_duration_minutes": {
            "type": "integer",
            "title": "Unit Duration Minutes",
            "description": "1ユニットの時間（分）: 10, 30, 60, 120"
          },
          "week_start_day": {
            "type": "string",
            "enum": [
              "monday",
              "sunday"
            ],
            "title": "Week Start Day",
            "description": "週開始曜日: monday or sunday"
          },
          "week_start_hour": {
            "type": "integer",
            "maximum": 23.0,
            "minimum": 0.0,
            "title": "Week Start Hour",
            "description": "週開始時刻（時）: 0-23"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時（ISO 8601）"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新日時（ISO 8601）"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "start_date",
          "end_date",
          "unit_duration_minutes",
          "week_start_day",
          "week_start_hour",
          "created_at",
          "updated_at"
        ],
        "title": "WeekResponse",
        "description": "週情報レスポンス."
      },
      "WeekUpdate": {
        "properties": {
          "unit_duration_minutes": {
            "type": "integer",
            "enum": [
              10,
              30,
              60,
              120
            ],
            "title": "Unit Duration Minutes",
            "description": "1ユニットの時間（分）: 10, 30, 60, 120"
          }
        },
        "type": "object",
        "required": [
          "unit_duration_minutes"
        ],
        "title": "WeekUpdate",
        "description": "週設定更新リクエスト."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}